+++ playground/nuxt.config.ts
@@ -2,12 +2,10 @@ export default defineNuxtConfig({
   modules: ['nuxt-log'],
   devtools: { enabled: true },
   runtimeConfig: {
-    public: {
-      logger: {
-        default: {
-          output: ['console'],
-          includeMeta: true,
-        },
+    logger: {
+      default: {
+        output: ['console'],
+        includeMeta: true,
       },
     },
   },